Reorganize model tests in single folder
