+++ python-client/tests/models/automodel/test_automodel.py
