+++ python-client/tests/models/automodel/test_infer_giskard_cls.py
