+++ python-client/tests/models/tensorflow/test_mnist.py
