+++ python-client/tests/models/tensorflow/test_tf_auto_model_as_embedding_layer.py
