+++ python-client/tests/models/test_model_auto_inference.py
